Update README.md
Hey! I appreciated the shout out 😄, but I'm not sure I feel comfortable sticking my name on this.

This is a community effort -- and I'm happy to play a role in organizing it, and I'll certainly stick it on a resume! But I don't want to put my name in the readme.